requirements.txt
streamlit
numpy
pickle5
tensorflow
scikit-learn
Pillow
